fix(docs): refactor nested layout component
